chinese_id/requirements.txt
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-image>=0.18.0
albumentations>=1.0.0
imgaug>=0.4.0
